cdppro/core/requirements.txt
## This file may be used to create an environment using:
## $ conda create --name <env> --file <this file>
## platform: win-64
aiohttp==3.5.4
mlxtend==0.17.0
gitdb2==2.0.5
gitpython==2.1.11
lime==0.1.1.33
mysql-connector-python==8.0.13
mysqlclient==1.3.14
numpy==1.16.3
pandas==0.24.2
python-crontab==2.3.6
requests==2.21.0
scikit-learn==0.20.3
sqlalchemy==1.3.3
